backend/requirements.txt
fastapi
uvicorn
motor
pymongo
httpx
google-genai
python-dotenv
pydantic
svix
clerk-backend-api
